--- controller/gardens/v1/gardens_test.go
@@ -1,1 +1,0 @@
-package gardens
