+++ routes/v1/routes.go
@@ -1,0 +1,11 @@
+package routes
+
+import (
+	"fmt"
+	"net/http"
+)
+
+func HomePage(w http.ResponseWriter, r *http.Request) {
+	fmt.Fprintf(w, "Welcome to the API Gateway!")
+	fmt.Println("Endpoint hit: homepage")
+}
